Split and move auth.ts to flows/authenticate
